requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
pandas>=1.5.0
openpyxl>=3.0.0 